+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-11-02
+
+### Added
+
 - Add label `"app.kubernetes.io/managed-by": "mia-platform"`
 - Unset original resource namespace
 - Add resource deletion if no longer deployed with `mlp`
@@ -29,7 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Release 🎉🎉🎉
 
-[Unreleased]: https://git.tools.mia-platform.eu/platform/devops/deploy/-/compare/v0.2.0...HEAD
+[Unreleased]: https://git.tools.mia-platform.eu/platform/devops/deploy/-/compare/v0.3.0...HEAD
+[0.3.0]: https://git.tools.mia-platform.eu/platform/devops/deploy/-/compare/v0.2.0...v0.3.0
 [0.2.0]: https://git.tools.mia-platform.eu/platform/devops/deploy/-/compare/v0.1.1...v0.2.0
 [0.1.1]: https://git.tools.mia-platform.eu/platform/devops/deploy/-/compare/v0.1.0...v0.1.1
 [0.1.0]: https://git.tools.mia-platform.eu/platform/devops/deploy/-/tags/v0.1.0